feat: move clustervariable test classes to its own package
